requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
markupsafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
tiktoken==0.5.2
requests==2.31.0
python-dotenv==1.0.0
flask-bootstrap==3.3.7.1
flask-wtf==1.2.1